Fix linker flags
There was an issue where a symbol that was created as an alias was
present in the final dylib but not the first pass dylib. Some
inspection of the circular library function found that the issue is
those flags were not having an effect on the first pass dylib
because they should be passed in as a parameter of the function.