Darling build of libdispatch-1173.60.1
Note: this *builds* but I'm not sure if it *runs*. Especially because of I had to enable a feature that was previously disabled in order to make it compile: kevent QoS.

P.S. to Apple: Would you mind making sure your code compiles for each "supported" configuration? Or otherwise just remove them entirely? Thanks.