Hack to fix manager root queue workers
This hack fixes crashes when the manager root queue is passed to a worker thread. (This was happening because of `_dispatch_kq_init` queuing up the manager thread).

Not sure if Apple's code was intentionally broken (i.e. maybe we're using an unsupported configuration) or if it was accidental, but either way, this most definitely feels like a hack (but it *does* prevent libdispatch from crashing, and I'm fairly limited in my debug capabilities at the moment to be able to tell if events are actually getting processed as they should).